Initial commit, moved some stuff around
