+++ seut_ot_exportBS.py
@@ -6,7 +6,7 @@ from .seut_ot_export                 import SEUT_OT_Export
 
 class SEUT_OT_ExportBS(Operator):
     """Exports Build Stages"""
-    bl_idname = "object.export_bs"
+    bl_idname = "object.export_buildstages"
     bl_label = "Export Build Stages"
     bl_options = {'REGISTER', 'UNDO'}
