+++ seut_pt_toolbar.py
@@ -74,7 +74,7 @@ class SEUT_PT_Panel_Export(bpy.types.Panel):
         col.operator('object.export_lod', text="LODs")
 
         col = split.column(align=True)
-        col.operator('object.export_bs', text="Build Stages")
+        col.operator('object.export_buildstages', text="Build Stages")
         col.operator('object.export_hkt', text="Collision")
 
         # Options
